monitor/requirements.txt
fastapi~=0.135.0
uvicorn[standard]~=0.42.0
pydantic-settings~=2.13.0
jinja2~=3.1.0
docker~=7.1.0
apscheduler~=3.11.0
httpx~=0.28.0
cryptography~=46.0.0
psycopg[binary]~=3.2.0
pyyaml~=6.0.0
